surface fix

+++ src/main/java/stefan/licenta/homefinder/dto/AdDto.java
@@ -13,7 +13,7 @@ public class AdDto {
     private String adType;
     private String description;
     private Double price;
-    private Integer surface;
+    private Double surface;
     private Integer rooms;
     private String adItemType;
     private Double lat;
